requirements.txt
aws-psycopg2>=1.3.8
boto3>=1.34.0
python-dotenv>=1.0.0